fix: moved profiles.js to /scripts
